Added AnimationRef to have a wrapper and checks to make sure animations are still valid when they are used. Moved extension methods to use AnimationRef

+++ src/Rust.UIFramework/Animation/AnimationId.cs
@@ -2,13 +2,10 @@
 
 namespace Oxide.Ext.UiFramework.Animation;
 
-public readonly record struct AnimationId(long Id)
+public readonly record struct AnimationId(int Id)
 {
     public bool IsValid => Id != 0;
-
-    private static long _nextAnimationId;
-    
+    private static int _nextAnimationId;
     internal static AnimationId GetNextId() => new(Interlocked.Increment(ref _nextAnimationId));
-    
     public override string ToString() => Id.ToString();
 }
